+++ NEWS.md
@@ -26,6 +26,10 @@
       Windows 7 or later.
   * Added more French translations from @NoNameWasDefined.
 
+* Bug fixes:
+  * Fix detection of small TGAs (under 64 bytes), e.g. some of the TGAs used by
+    Steam for its user interface.
+
 ## v2.6 (released 2025/09/13)
 
 * New parsers: